requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
Augur==0.1
Authlib==1.3.2
backoff==2.2.1
bandit==1.8.0
beautifulsoup4==4.12.3
betterproto==2.0.0b6
boolean.py==4.0
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cyclonedx-python-lib==7.6.2
defusedxml==0.7.1
Deprecated==1.2.15
dnspython==2.7.0
dparse==0.6.4
email_validator==2.2.0
filelock==3.12.4
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.5
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
id==1.4.0
idna==3.10
Jinja2==3.1.4
license-expression==30.4.0
mailbits==0.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
ort==1.2.1
ossaudit==0.5.0
packageurl-python==0.16.0
packaging==24.2
pbr==6.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
platformdirs==4.3.6
propcache==0.2.1
psutil==6.0.0
py-serializable==1.1.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pypi-simple==1.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rfc8785==0.1.4
rich==13.9.4
ruamel.yaml==0.18.6
safety==3.2.11
safety-schemas==0.0.10
securesystemslib==1.1.0
setuptools==75.6.0
shellingham==1.5.4
sigstore==3.5.3
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stevedore==5.4.0
texttable==1.7.0
toml==0.10.2
tuf==5.1.0
typer==0.14.0
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.3
